+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.shell.interactive.enabled=true
 spring.application.name=HyperMetro
 application.title=HyperMetro
 application.version=2.0
+hypermetro.transfer.time=5
